Extract Pipeline class
